requirements.txt
flet==0.24.1
pillow==10.4.0
pyinstaller==6.10.0
pytest==8.3.3
coverage==7.6.1
ruff==0.6.4
black==24.8.0
